creates padPokemonId helper function
