+++ src/components/Pokemon/Sprites/index.tsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 // types
 import type { Pokemon, PokemonSpecies, PokemonSprites } from 'pokenode-ts';
 // helpers
-import { removeUnderscore } from '@/helpers/typography';
+import { removeUnderscore, padPokemonId } from '@/helpers';
 // styles
 import { SectionTitle, SectionSubTitle } from '@/components/BaseStyles';
 import { SpriteContainer, Sprite, SpriteSubtitle, NoSprites } from './StyledSprites';
@@ -154,9 +154,9 @@ const Sprites = ({ pokemonSprites, pokemonId, forms, ...rest }: SpritesProps): J
                     <Sprite
                       alt="Official Artwork Front Default"
                       key="official-artwork"
-                      src={`https://raw.githubusercontent.com/andreferreiradlw/pokestats_media/main/assets/images/${pokemonId
-                        .toString()
-                        .padStart(3, '0')}-${name}.png`}
+                      src={`https://raw.githubusercontent.com/andreferreiradlw/pokestats_media/main/assets/images/${padPokemonId(
+                        pokemonId,
+                      )}-${name}.png`}
                       height="180"
                     />
                   </SpriteContainer>
