+++ src/helpers/typography.ts
@@ -4,4 +4,14 @@ const removeUnderscore = (str: string): string => str.replace(/_/g, ' ');
 // remove dashes
 const removeDash = (str: string): string => str.replace(/-/g, ' ');
 
-export { removeUnderscore, removeDash };
+const formatFlavorText = (text: string): string =>
+  text
+    .replace(/\u00AD/g, '')
+    .replace(/\u000C/g, ' ')
+    .replace(/u' -\n'/, ' - ')
+    .replace(/u'-\n'/, '-')
+    .replace(/(\r\n|\n|\r)/gm, ' ');
+
+const padPokemonId = (id: number): string => id.toString().padStart(3, '0');
+
+export { removeUnderscore, removeDash, formatFlavorText, padPokemonId };
